@@ -3,29 +3,31 @@ import { BrowserModule } from '@angular/platform-browser';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { HeaderComponent } from './component/header/header.component';
-import { HomeComponent } from './pages/home/home.component';
-import { NotFoundComponent } from './pages/not-found/not-found.component';
-import { MenuComponent } from './pages/menu/menu.component';
-import { LoginComponent } from './pages/login/login.component';
-import { MatSlideToggleModule } from '@angular/material/slide-toggle';
+import { HomeComponent } from './component/home/home.component';
+import { NotFoundComponent } from './component/not-found/not-found.component';
+import { MenuComponent } from './component/menu/menu.component';
+import { LoginComponent } from './component/auth/login/login.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { MatFormFieldModule } from '@angular/material/form-field';
 import { ReactiveFormsModule } from '@angular/forms';
-import { MatIconModule } from '@angular/material/icon';
-import { MatExpansionModule } from '@angular/material/expansion';
-import { ProductComponent } from './component/product/product.component';
-import { MatToolbarModule } from '@angular/material/toolbar';
-import { MatMenuModule } from '@angular/material/menu';
+import { ProductComponent } from './component/menu/product/product.component';
 import { FormsModule } from '@angular/forms';
-import { RegisterComponent } from './pages/register/register.component';
+import { RegisterComponent } from './component/auth/register/register.component';
 import { HttpClientModule } from '@angular/common/http';
-import { ShoppingCartComponent } from './component/shopping-cart/shopping-cart.component';
-import { HamburgerMenuComponent } from './component/hamburger-menu/hamburger-menu.component';
-import { ProductCartComponent } from './component/product-cart/product-cart.component'; // Necesario para habilitar las solicitudes HTTP
-import { RegisteredProfileComponent } from './pages/profiles/registered-profile/registered-profile.component';
-import { UnregisteredProfileComponent } from './pages/profiles/unregistered-profile/unregistered-profile.component';
-import { RightMenuComponent } from './component/right-menu/right-menu.component';
-import { LeftMenuComponent } from './component/left-menu/left-menu.component';
+import { ShoppingCartComponent } from './component/header/right-menu/shopping-cart/shopping-cart.component';
+import { HamburgerMenuComponent } from './component/header/left-menu/hamburger-menu/hamburger-menu.component';
+import { ProductCartComponent } from './component/header/right-menu/shopping-cart/product-cart/product-cart.component';
+import { RightMenuComponent } from './component/header/right-menu/right-menu.component';
+import { LeftMenuComponent } from './component/header/left-menu/left-menu.component';
+import { MyOrdersComponent } from './component/header/left-menu/my-orders/my-orders.component';
+import { OrderComponent } from './component/header/left-menu/my-orders/order/order.component';
+import { CouponsComponent } from './component/header/left-menu/coupons/coupons.component';
+import { CuponComponent } from './component/header/left-menu/coupons/cupon/cupon.component';
+import { ProfileComponent } from './component/header/left-menu/profile/profile.component';
+import { DiscountComponent } from './component/header/left-menu/discount/discount.component';
+import { BackofficeComponent } from './component/backoffice/backoffice.component';
+import { CategoryBOComponent } from './component/backoffice/category-bo/category-bo.component';
+import { ProductoBOComponent } from './component/backoffice/producto-bo/producto-bo.component';
+import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 
 @NgModule({
   declarations: [
@@ -40,24 +42,26 @@ import { LeftMenuComponent } from './component/left-menu/left-menu.component';
     ShoppingCartComponent,
     HamburgerMenuComponent,
     ProductCartComponent,
-    RegisteredProfileComponent,
-    UnregisteredProfileComponent,
     RightMenuComponent,
-    LeftMenuComponent
+    LeftMenuComponent,
+    MyOrdersComponent,
+    OrderComponent,
+    CouponsComponent,
+    CuponComponent,
+    ProfileComponent,
+    DiscountComponent,
+    BackofficeComponent,
+    CategoryBOComponent,
+    ProductoBOComponent
   ],
   imports: [
     BrowserModule,
     AppRoutingModule,
-    MatSlideToggleModule,
     BrowserAnimationsModule,
-    MatFormFieldModule,
     ReactiveFormsModule,
-    MatToolbarModule,
-    MatMenuModule,
-    MatIconModule,
-    MatExpansionModule,
     FormsModule,
-    HttpClientModule
+    HttpClientModule,
+    NgbModule
   ],
   providers: [],
   bootstrap: [AppComponent]
