@@ -1,10 +1,11 @@
-import { Detail } from "./detail";
-
 export interface Product {
-  name: string;
-  description: string;
-  price: number;
-  image: string;
-  details: Detail[];
-  discount: number;
+  id: number
+  name: string
+  description: string
+  image: string
+  price: number
+  stock: number
 }
+
+
+
